+++ package.json
@@ -10,16 +10,22 @@
     "format": "prettier --write src/"
   },
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.3.0",
+    "@fortawesome/free-brands-svg-icons": "^6.3.0",
+    "@fortawesome/vue-fontawesome": "^3.0.3",
     "@supabase/supabase-js": "^2.10.0",
     "dayjs": "^1.11.7",
+    "simple-icons": "^8.8.0",
     "v-lazy-image": "^2.1.1",
     "vue": "^3.2.47",
+    "vue-3-linkify": "^1.1.0",
     "vue-auth-ui": "^0.1.1",
     "vue-router": "^4.1.6",
     "vue3-dayjs-plugin": "^1.0.2",
     "vuejs-medium-editor": "^3.0.2"
   },
   "devDependencies": {
+    "@fortawesome/fontawesome-free": "^6.3.0",
     "@rushstack/eslint-patch": "^1.2.0",
     "@vitejs/plugin-vue": "^4.0.0",
     "@vue/eslint-config-prettier": "^7.1.0",
